♻️ Move list-example module to list-example folder
