+++ apps/cookbook/src/app/examples/examples.module.ts
@@ -14,7 +14,7 @@ import { COMPONENT_DECLARATIONS } from './examples.common';
 import { FlagExampleModule } from './flag-example/flag-example.module';
 import { FormFieldExamplesModule } from './form-field-example/form-field-example.module';
 import { ItemExampleModule } from './item-example/item-example.module';
-import { ListExamplesModule } from './list/list-example.module';
+import { ListExamplesModule } from './list-example/list-example.module';
 import { ModalCompactExampleComponent } from './modal-example/compact-example/modal-compact-example.component';
 import { FirstEmbeddedModalExampleComponent } from './modal-example/first-embedded-modal-example/first-embedded-modal-example.component';
 import { ModalRoutePage1ExampleComponent } from './modal-example/modal-route-example/modal-route-page1-example.component';
