+++ apps/cookbook/src/app/examples/examples.routes.ts
@@ -22,18 +22,18 @@ import { FormFieldExampleComponent } from './form-field-example/form-field-examp
 import { GridExampleComponent } from './grid-example/grid-example.component';
 import { IconExampleComponent } from './icon-example/icon-example.component';
 import { ItemExampleComponent } from './item-example/item-example.component';
+import { ListBoldTextOnRowSelectionExampleComponent } from './list-example/components/bold-text-on-row-selection';
+import { ListColoredItemsExampleComponent } from './list-example/components/colored-items';
+import { ListWithDividersExampleComponent } from './list-example/components/dividers';
+import { ListWithHeaderAndFooterExampleComponent } from './list-example/components/header-and-footer';
+import { ListItemsExampleComponent } from './list-example/components/items';
+import { ListWithSectionsExampleComponent } from './list-example/components/sections';
+import { ListWithSectionsAndColoredItemsExampleComponent } from './list-example/components/sections-and-colored-items';
+import { ListSelectableItemsExampleComponent } from './list-example/components/selectable-items';
 import { ListExampleComponent } from './list-example/list-example.component';
 import { ListLoadOnDemandExampleComponent } from './list-load-on-demand-example/list-load-on-demand-example.component';
 import { ListNoShapeExampleComponent } from './list-no-shape-example/list-no-shape-example.component';
 import { ListSwipeExampleComponent } from './list-swipe-example/list-swipe-example.component';
-import { ListBoldTextOnRowSelectionExampleComponent } from './list/components/bold-text-on-row-selection';
-import { ListColoredItemsExampleComponent } from './list/components/colored-items';
-import { ListWithDividersExampleComponent } from './list/components/dividers';
-import { ListWithHeaderAndFooterExampleComponent } from './list/components/header-and-footer';
-import { ListItemsExampleComponent } from './list/components/items';
-import { ListWithSectionsExampleComponent } from './list/components/sections';
-import { ListWithSectionsAndColoredItemsExampleComponent } from './list/components/sections-and-colored-items';
-import { ListSelectableItemsExampleComponent } from './list/components/selectable-items';
 import { LoadingOverlayExampleComponent } from './loading-overlay-example/loading-overlay-example.component';
 import { ModalExampleComponent } from './modal-example/modal-example.component';
 import { ModalRoutePage1ExampleComponent } from './modal-example/modal-route-example/modal-route-page1-example.component';
