+++ apps/cookbook/src/app/examples/list-example/components/bold-text-on-row-selection.ts
