+++ apps/cookbook/src/app/examples/list-example/components/colored-items.ts
