+++ apps/cookbook/src/app/examples/list-example/components/items.ts
