+++ apps/cookbook/src/app/examples/list-example/components/sections-and-colored-items.ts
