+++ apps/cookbook/src/app/examples/list-example/components/selectable-items.ts
